refactor: Move setup files to config package
